Magic sets impl compare to exhaustive
